Decoupling some methods and changes package structure following review by Javier
